requirements.txt
requests
django
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
cloudinary
moviepy

langchain-text-splitters 
langchain-community 
langchain-openai
langchain-postgres
langchain-core
langgraph